+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 ## [1.0.0] 2023-10-22
 
 - Initial release
+
+## [1.0.3] 2023-10-25
+
+- Add highlighting
+
+## [1.0.5] 2023-10-26
+
+- Add ignore config
+- Fix webview fails to load after reloading vs code